refactor: isolate session sharing management actions
